backend/pyproject.toml
[project]
name = "vine-invest-research"
version = "0.1.0"
description = "Investment Research AI - Full-stack multi-agent system"
requires-python = ">=3.11,<3.13"

dependencies = [
    "google-adk>=1.0.0",
    "litellm>=1.40.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "PyMuPDF>=1.24.0",
    "tavily-python>=0.5.0",
    "python-multipart>=0.0.9",
    "sse-starlette>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
